Work on displaying project on screen
Added code that displays the rest of the content for the project
description. Still need to mae some adjustments to it however.

Also added a function that should remove the displayed project,
and display the selected project from the sidebar. I still need to
couple this one with an event listener which I plan to add to the
sidebar buttons..